+++ packages/core-js/internals/iterator-map.js
@@ -1,10 +1,13 @@
 'use strict';
 var call = require('../internals/function-call');
-var aCallable = require('../internals/a-callable');
 var anObject = require('../internals/an-object');
 var getIteratorDirect = require('../internals/get-iterator-direct');
 var createIteratorProxy = require('../internals/iterator-create-proxy');
 var callWithSafeIterationClosing = require('../internals/call-with-safe-iteration-closing');
+var iteratorClose = require('../internals/iterator-close');
+var globalThis = require('../internals/global-this');
+var checkIteratorClosingOnEarlyError = require('../internals/check-iterator-closing-on-early-error');
+var aCallable = require('./a-callable');
 
 var IteratorProxy = createIteratorProxy(function () {
   var iterator = this.iterator;
@@ -13,11 +16,21 @@ var IteratorProxy = createIteratorProxy(function () {
   if (!done) return callWithSafeIterationClosing(iterator, this.mapper, [result.value, this.counter++], true);
 });
 
+var Iterator = globalThis.Iterator;
+var nativeMap = Iterator && Iterator.prototype && Iterator.prototype.map;
+
 // `Iterator.prototype.map` method
 // https://github.com/tc39/proposal-iterator-helpers
-module.exports = function map(mapper) {
+module.exports = (nativeMap && checkIteratorClosingOnEarlyError(TypeError, nativeMap, null)) ? nativeMap : function map(mapper) {
   anObject(this);
-  aCallable(mapper);
+  try {
+    aCallable(mapper);
+  } catch (error) {
+    iteratorClose(this, 'throw', error);
+  }
+
+  if (nativeMap) return call(nativeMap, this, mapper);
+
   return new IteratorProxy(getIteratorDirect(this), {
     mapper: mapper
   });