+++ packages/core-js/modules/es.iterator.filter.js
@@ -1,12 +1,15 @@
 'use strict';
 var $ = require('../internals/export');
 var call = require('../internals/function-call');
-var aCallable = require('../internals/a-callable');
 var anObject = require('../internals/an-object');
 var getIteratorDirect = require('../internals/get-iterator-direct');
 var createIteratorProxy = require('../internals/iterator-create-proxy');
 var callWithSafeIterationClosing = require('../internals/call-with-safe-iteration-closing');
 var IS_PURE = require('../internals/is-pure');
+var iteratorClose = require('../internals/iterator-close');
+var globalThis = require('../internals/global-this');
+var checkIteratorClosingOnEarlyError = require('../internals/check-iterator-closing-on-early-error');
+var aCallable = require('../internals/a-callable');
 
 var IteratorProxy = createIteratorProxy(function () {
   var iterator = this.iterator;
@@ -22,12 +25,24 @@ var IteratorProxy = createIteratorProxy(function () {
   }
 });
 
+var Iterator = globalThis.Iterator;
+var nativeFilter = Iterator && Iterator.prototype && Iterator.prototype.filter;
+var NATIVE_METHOD_WITHOUT_CLOSING_ON_EARLY_ERROR = nativeFilter && !checkIteratorClosingOnEarlyError(TypeError, nativeFilter, null);
+var FORCED = IS_PURE || !nativeFilter || NATIVE_METHOD_WITHOUT_CLOSING_ON_EARLY_ERROR;
+
 // `Iterator.prototype.filter` method
 // https://tc39.es/ecma262/#sec-iterator.prototype.filter
-$({ target: 'Iterator', proto: true, real: true, forced: IS_PURE }, {
+$({ target: 'Iterator', proto: true, real: true, forced: FORCED }, {
   filter: function filter(predicate) {
     anObject(this);
-    aCallable(predicate);
+    try {
+      aCallable(predicate);
+    } catch (error) {
+      iteratorClose(this, 'throw', error);
+    }
+
+    if (NATIVE_METHOD_WITHOUT_CLOSING_ON_EARLY_ERROR) return call(nativeFilter, this, predicate);
+
     return new IteratorProxy(getIteratorDirect(this), {
       predicate: predicate
     });