+++ packages/core-js/modules/es.iterator.flat-map.js
@@ -1,13 +1,15 @@
 'use strict';
 var $ = require('../internals/export');
 var call = require('../internals/function-call');
-var aCallable = require('../internals/a-callable');
 var anObject = require('../internals/an-object');
 var getIteratorDirect = require('../internals/get-iterator-direct');
 var getIteratorFlattenable = require('../internals/get-iterator-flattenable');
 var createIteratorProxy = require('../internals/iterator-create-proxy');
 var iteratorClose = require('../internals/iterator-close');
 var IS_PURE = require('../internals/is-pure');
+var globalThis = require('../internals/global-this');
+var checkIteratorClosingOnEarlyError = require('../internals/check-iterator-closing-on-early-error');
+var aCallable = require('../internals/a-callable');
 
 var IteratorProxy = createIteratorProxy(function () {
   var iterator = this.iterator;
@@ -31,12 +33,24 @@ var IteratorProxy = createIteratorProxy(function () {
   }
 });
 
+var Iterator = globalThis.Iterator;
+var nativeFlatMap = Iterator && Iterator.prototype && Iterator.prototype.flatMap;
+var NATIVE_METHOD_WITHOUT_CLOSING_ON_EARLY_ERROR = nativeFlatMap && !checkIteratorClosingOnEarlyError(TypeError, nativeFlatMap, null);
+var FORCED = IS_PURE || !nativeFlatMap || NATIVE_METHOD_WITHOUT_CLOSING_ON_EARLY_ERROR;
+
 // `Iterator.prototype.flatMap` method
 // https://tc39.es/ecma262/#sec-iterator.prototype.flatmap
-$({ target: 'Iterator', proto: true, real: true, forced: IS_PURE }, {
+$({ target: 'Iterator', proto: true, real: true, forced: FORCED }, {
   flatMap: function flatMap(mapper) {
     anObject(this);
-    aCallable(mapper);
+    try {
+      aCallable(mapper);
+    } catch (error) {
+      iteratorClose(this, 'throw', error);
+    }
+
+    if (NATIVE_METHOD_WITHOUT_CLOSING_ON_EARLY_ERROR) return call(nativeFlatMap, this, mapper);
+
     return new IteratorProxy(getIteratorDirect(this), {
       mapper: mapper,
       inner: null