+++ tests/unit-pure/es.iterator.every.js
@@ -28,5 +28,7 @@ QUnit.test('Iterator#every', assert => {
   assert.throws(() => every.call([], () => { /* empty */ }), TypeError);
   assert.throws(() => every.call(createIterator([1]), undefined), TypeError);
   assert.throws(() => every.call(createIterator([1]), null), TypeError);
-  assert.throws(() => every.call(createIterator([1]), {}), TypeError);
+  const it = createIterator([1], { return() { this.closed = true; } });
+  assert.throws(() => every.call(it, {}), TypeError);
+  assert.true(it.closed, "every doesn't close iterator on validation error");
 });